Reorganized files and created a data_loader.py file
