+++ examples/linear_regression.py
@@ -1,67 +1,11 @@
-from typing import Tuple
-
-import numpy as np
-import pandas as pd
 from showml.losses import MeanSquareError
 from showml.optimizers import SGD
 from showml.supervised.regression import LinearRegression
 from showml.utils.dataset import Dataset
-from showml.utils.metrics import mean_square_error, r2_score
+from showml.losses.metrics import mean_square_error, r2_score
 from showml.utils.plots import plot_regression_line
 from showml.utils.preprocessing import normalize
-
-
-def load_auto() -> Tuple[np.ndarray, np.ndarray]:
-    """
-    A method to load the Auto.csv file, include the necessary columns and return X_train and y_train
-    return X_train: The input training data
-    return y_train: The input training labels
-    """
-    auto = (
-        pd.read_csv("./data/Auto.csv", na_values="?", dtype={"ID": str})
-        .dropna()
-        .reset_index()
-    )
-    X_train = auto[
-        [
-            "cylinders",
-            "displacement",
-            "horsepower",
-            "weight",
-            "acceleration",
-            "year",
-            "origin",
-        ]
-    ].values
-    y_train = auto[["mpg"]].values
-
-    # Make y_train 1D if its not
-    if y_train.ndim > 1:
-        y_train = y_train[:, 0]
-
-    return X_train, y_train
-
-
-def load_salary() -> Tuple[np.ndarray, np.ndarray]:
-    """
-    A method to load the Salary.csv file, include the necessary columns and return X_train and y_train
-    return X_train: The input training data
-    return y_train: The input training labels
-    """
-    salary_data = (
-        pd.read_csv("./data/Salary.csv", na_values="?", dtype={"ID": str})
-        .dropna()
-        .reset_index()
-    )
-    X_train = salary_data[["YearsExperience"]].values
-    y_train = salary_data[["Salary"]].values
-
-    # Make y_train 1D if its not
-    if y_train.ndim > 1:
-        y_train = y_train[:, 0]
-
-    return X_train, y_train
-
+from showml.utils.data_loader import load_salary, load_auto
 
 X_train, y_train = load_auto()
 X_train = normalize(X_train)
